Add $$GMIC_LIB_PATH qmake variable
Add a GMIC_LIB_PATH variable to the qmake build system, so it is
possible to specify where to find the dynamic gmic library, in the
same way it is possible in gmic-qt. This way, one can build gmic with
builddir!=srcdir, and build zart using libgmic (which would then be in
a different location than the gmic sources).

If not specified, it falls back to $$GMIC_PATH, so behaving the same
as before.