Add $$GMIC_LIB_PATH for zart (Patch by Pino Toscano)
Author: Pino Toscano <pino@debian.org>
Description: Add $$GMIC_LIB_PATH for zart
Last-Update: 2018-11-10

Add a GMIC_LIB_PATH variable to the qmake build system of zart, so it  is
possible to specify where to find the dynamic gmic library, in the  same way it
is possible in gmic-qt.

If not specified, it falls back to $$GMIC_PATH, so behaving the same  as
before.